Move type traits that select storage implementation to storage file
